Changed variables & made new concepts
The new main variable is now "startingMode" which is changed at the start only by going up and down. Currently power must be restarted to restart the clock, and there is a lot of now pointless code that tries to account for restarts, I don't care. I'm sure debugging needs to be done so that "...Base" and "...MilliSeconds" begin interacting again correctly, however right now the console gives no errors.